serveur/chat_bot/requirements.txt
flask
psycopg2-binary
ollama
langchain
langchain-community
langchain-ollama
langchain-text-splitters
chromadb